error handling for when a thing does not exist

+++ ration/api/controllers/things/destroy-one-thing.js
@@ -11,6 +11,11 @@ module.exports = {
     },
 
     exits: {
+        notFound: {
+            description:
+                'The thing the users is trying to delete does not exist',
+            responseType: 'notFound'
+        },
         forbidden: {
             description:
                 'The users making this request does not have the permissions to delete this thing',
@@ -21,12 +26,17 @@ module.exports = {
     fn: async function(inputs) {
         const thing = await Thing.findOne({ id: inputs.id });
 
+        // if the thing trying to delete does not exists
+        if (!thing) {
+            throw 'notFound';
+        }
+
         if (thing.owner !== this.req.me.id) {
             throw 'forbidden';
         }
 
         await Thing.destroy({ id: inputs.id });
 
-        return;
+        return
     }
 };
